shared/requirements.txt
# Shared Module Dependencies

# Configuration parsing
pyyaml>=6.0

# Database
psycopg2-binary>=2.9.0

# Data validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client for service checking
httpx>=0.25.0
